DjangoWebCrawler/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
Deprecated==1.2.13
Django==4.0.2
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
packaging==21.3
pyparsing==3.0.7
pytz==2021.3
redis==4.1.3
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
wrapt==1.13.3
